create models for Libro and Autor

+++ src/main/java/com/oraclenextone/aluracursos/literalura/modelos/DatosLibro.java
@@ -10,6 +10,6 @@ public record DatosLibro(
         @JsonAlias("title") String titulo,
         @JsonAlias("authors") List<DatosAutor> autor,
         @JsonAlias("languages") List<String> idiomas,
-        @JsonAlias("download_count") Double descargas
+        @JsonAlias("download_count") Integer descargas
 ) {
 }
